implemented StoredPlayer Object
Player objects are crated/loaded on join to reduce the load on the server. this should also be dont for the server, but cant atm due to how the hasServer checks are done. For the server and player files, they have been moved to the data folder to split up instances for less resource usage. from this point on i should no longer have to make config breaking changes. And yes this was needed. Major api changes have been made so expect addons to break

Update PlayerServers.java